requirements.txt
fastapi>=0.135.2
uvicorn[standard]>=0.42.0
pydantic>=2.0
python-dotenv>=1.0
supabase>=2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
openpyxl>=3.1
rarfile>=4.1
py7zr>=0.21
notebooklm-py>=0.3.4
cryptography>=44.0
